refactory OpenApi configuration
