+++ otp-acl/src/model/mod.rs
@@ -1,2 +1,3 @@
 pub mod otp_keys;
 pub mod schema;
+pub mod api_doc;
